requirements.txt
assertpy==1.1
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
exceptiongroup==1.1.1
execnet==1.9.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
outcome==1.2.0
packaging==23.1
pluggy==1.2.0
py==1.11.0
PyHamcrest==2.0.4
PySocks==1.7.1
pytest==7.4.0
pytest-html==3.2.0
pytest-metadata==3.0.0
pytest-xdist==3.3.1
python-dotenv==1.0.0
selenium==4.10.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.3
urllib3==2.0.3
wsproto==1.2.0
